refactor(element-template): move subcommand handlers into commands/
Subcommand entry points (apply, get, get-properties, info, search, sync)
now live under default-plugins/element-template/commands/. The plugin
root keeps c8ctl-plugin.ts (dispatch), template-ref.ts, helpers.ts,
marketplace.ts — the shared utilities are no longer mixed with
command modules.

Updates relative import paths in the moved files (../helpers.ts etc.)
and bumps the vendor-bundle path in commands/apply.ts by one level.
Refreshes the file table in AGENTS.md.
